purge sphinx context manager
This was a neat idea, but unfortunately upstream sphinx keeps multiple
references to itself and internal structures, making a web of variables
that do not leave the execution context once loaded.

This causes a litany of parsing errors that have only gotten worse as
sphinx progresses. For now, load one instance of sphinx to interact with
if sphinx is available on the system.

Signed-off-by: Randolph Sapp <rs@ti.com>